Deserialisation done
Thanks to Tamriel Trade Centre, I have a simple library for Lua
deserialisation. Implemented and tested working.
Application will now correctly display all research times for the
selected character (which defaults to the first character on the first
account listed in the SavedVariables) and show notifications when it's
done.
Next step: Saving selections made in the UI (account & character as well
as notification appearance settings).