Tooltips!
Added tooltips for all of the UI, mainly because long item/trait/time
combinations would be cut off more often than not, so the tooltip for
that label echoes the label text; other tooltips provide a short
explanation for the UI element. Sometimes redundantly so where the UI is
self-explanatory.